be/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.8.0
argcomplete==3.5.3
asttokens==2.4.1
astunparse==1.6.3
attrs==25.1.0
bcrypt==4.2.1
black==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
datamodel-code-generator==0.26.5
debugpy==1.8.5
decorator==5.1.1
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
eval_type_backport==0.2.2
executing==2.1.0
filelock==3.17.0
flasgger==0.9.7.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-PyMongo==3.0.1
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
flatbuffers==24.12.23
frozenlist==1.5.0
function-calling==0.0.1
gast==0.6.0
genson==1.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
google-pasta==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
groq==0.18.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
inflect==5.6.2
ipykernel==6.29.5
ipython==8.27.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.5
jiter==0.8.2
jose==1.0.0
jpype1==1.5.1
jsonpath-python==1.0.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.7.0
konlpy==0.6.0
libclang==18.1.1
llm-opt==0.1.4
llm-taxi==0.7.0
lxml==5.3.0
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistralai==1.5.0
mistune==3.1.1
ml-dtypes==0.4.1
mongoengine==0.29.1
multidict==6.1.0
mypy-extensions==1.0.0
mysql-connector-python==9.1.0
namex==0.0.8
nest-asyncio==1.6.0
numpy==2.0.2
openai==1.61.0
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
parso==0.8.4
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.2
prompt_toolkit==3.0.47
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.2
psutil==6.0.0
pure_eval==0.2.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
pymongo==4.10.1
PyMySQL==1.1.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2025.1
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
setuptools==75.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
SQLAlchemy==2.0.36
stack-data==0.6.3
structlog==25.1.0
tabulate==0.9.0
tavily-python==0.5.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tiktoken==0.8.0
together==1.4.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.44.0
wrapt==1.17.0
yarl==1.18.3
